requirements.txt
langchain
langchain-openai
python-dotenv
pinecone-client
streamlit
openai
unstructured[pdf]
